# docs/openapi/ferrum.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ferrum",
    "description": "Implementation map of the Ferrum gateway (utoipa dump). Source of truth for GA4GH DRS/WES/TES/TRS/Beacon/htsget is each standard’s published OpenAPI — see docs/GA4GH.md. This file records what this process exposes, including Ferrum-only paths. Not a replacement spec. Not official GA4GH certification. Auth is a deployment choice. Behavioral proof against the published specs is HelixTest.",
    "contact": {
      "name": "Synaptic Four",
      "email": "contact@synapticfour.com"
    },
    "license": {
      "name": "BUSL-1.1",
      "url": "https://github.com/SynapticFour/Ferrum/blob/main/LICENSE"
    },
    "version": "0.3.1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Paths are absolute from the Ferrum gateway origin (default http://127.0.0.1:8080)."
    }
  ],
  "paths": {
...
    "/ga4gh/trs/v2/internal/register": {
      "post": {
        "tags": [
          "handlers"
        ],
        "summary": "Internal: register a tool (e.g. from WES when a workflow is submitted).",
        "operationId": "register_tool",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          }
        }
      }
    },
...
    "/ga4gh/trs/v2/tools": {
      "get": {
        "tags": [
          "handlers"
        ],
        "summary": "GET /tools returns a root-level JSON array for GA4GH/HelixTest compatibility (\"TRS /tools must return array\").",
        "operationId": "list_tools",
        "parameters": [
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tool"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ga4gh/trs/v2/tools/{id}": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_tool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/ga4gh/trs/v2/tools/{id}/versions": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_tool_versions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolVersion"
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/ga4gh/trs/v2/tools/{id}/versions/{version_id}/descriptor/{descriptor_type}": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_descriptor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "descriptor_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterToolRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "organization": {
            "type": "string",
            "nullable": true
          },
          "toolclass": {
            "type": "string",
            "nullable": true
          },
          "workflow_content": {
            "type": "string",
            "description": "Inline workflow file body (alternative to `workflow_url`).",
            "nullable": true
          },
          "workflow_type": {
            "type": "string",
            "nullable": true
          },
          "workflow_type_version": {
            "type": "string",
            "nullable": true
          },
          "workflow_url": {
            "type": "string",
            "description": "Remote descriptor URL (WDL/CWL/NFL/SMK). Required unless `workflow_content` is set.",
            "nullable": true
          }
        }
      },
...
      "Tool": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "meta_version": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "organization": {
            "type": "string",
            "nullable": true
          },
          "toolclass": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolClass"
              }
            ],
            "nullable": true
          },
          "url": {
            "type": "string",
            "description": "GA4GH TRS: URL to fetch this tool (e.g. /ga4gh/trs/v2/tools/{id}).",
            "nullable": true
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolVersion"
            },
            "description": "GA4GH TRS: versions list (populated in list response for schema compliance).",
            "nullable": true
          }
        }
      },
      "ToolClass": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ToolVersion": {
        "type": "object",
        "required": [
          "id",
          "name",
          "tool_id",
          "url"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tool_id": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "GA4GH TRS: URL to fetch this tool version (e.g. /ga4gh/trs/v2/tools/{id}/versions)."
          }
        }
      },
...
    }
  }
}